+++ lld/docs/ReleaseNotes.rst
@@ -25,6 +25,11 @@ Non-comprehensive list of changes in this release
 
 ELF Improvements
 ----------------
+* For AArch64, support for the ``-zgcs-report-dynamic`` option has been added. This will provide users with
+  the ability to check any Dynamic Objects explicitly passed to LLD for the GNU GCS Attribute Flag. This is
+  required for all files when linking with GCS enabled. Unless defined by the user, ``-zgcs-report-dynamic``
+  inherits its value from the ``-zgcs-report`` option, capped at the ``warning`` level to ensure that a users
+  module can still compile. This behaviour is designed to match the GNU ld Linker.
 
 Breaking changes
 ----------------